image-colorization/Task 2/requirements.txt
torch
torchvision
torchaudio
--index-url https://download.pytorch.org/whl/cu124
matplotlib
numpy
Pillow
opencv-python
tk
